requirements.txt
anyio==3.5.0
asgiref==3.5.1
asyncpg==0.25.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
databases==0.5.5
ecdsa==0.17.0
fastapi==0.77.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
install==1.3.5
passlib==1.7.4
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
